Fixed the difficulty
Split the move lists into two, a set to directly block a win and a set
for a strategic play to get a corner. The second set also further splits
into 3 patterns that get rotated.

The losits have been reordered to make it easier to see gaps or breaks
in the patterns.

Added moves that rely on playing for the centre, while technically
redundant having the patterns be regular makes it easier to see if a
mistake gets made